% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disp_vel_lin}
\alias{disp_vel_lin}
\title{Function to calculate displaement and velocity of a projectile when air friction is introduced
This function only treats the linear case of velocity dependence.}
\usage{
disp_vel_lin(y0, v0, theta0, t, b, m)
}
\arguments{
\item{v0}{initial velocity in m/s}

\item{y0}{initial height in m}

\item{theta0}{initial angle in degrees}

\item{t}{Time of flight in seconds}

\item{b}{drag coefficient in Newtons.seconds/meters}

\item{m}{mass of object in kg}
}
\value{
x Displacement in the horizontal direction in meters

vx speed in the horizontal direction at a given time t (m/s units)

y Displacement in the vertical direction in meters

vy speed in the vertical direction at a given time t (m/s units)
}
\description{
This function takes in the initial velocity, time of flight,
drag coefficient, and mass of object and returns the distance
travlled and speed at a given time t
}

