% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findvar}
\alias{findvar}
\title{Find column name corresponding to a particular functional}
\usage{
findvar(x, v, ignorecase = TRUE)
}
\arguments{
\item{x}{The data we are querying (data.frame)}

\item{v}{The keyword we are searching for (character)}

\item{ignorecase}{Should case be ignored (logical)}
}
\value{
A vector of column names matching the keyword
}
\description{
The original data set contains very long column headers. This function
does a keyword search over the headers to find those column headers that
match a particular keyword, e.g., mean, median, etc.
}

