% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jit}
\alias{jit}
\title{jit}
\usage{
jit(x, n, amount = NULL, method = "norm")
}
\arguments{
\item{x}{a numerical Vector}

\item{n}{an integer representing how many times to repeat the jittering
for each entry in x. If n is a non-integer numeric, it will coerced
into an integer}

\item{amount}{same as in \link{jitter} when using unifrom values
and is equal to the standard deviation when using normal values}

\item{method}{randomizing method: "norm" or "uniform"}
}
\description{
simple function to jitter a vector by the specified amount
after repeating it n numbers
}
\examples{
jit(x = 40, n = 10, method = "norm", amount=3)
jit(x=10, n=5, method="uniform")
}

