% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Projectile}
\alias{Projectile}
\title{Function to calculate displaement and velocity of a projectile in vacuum}
\usage{
Projectile(t, y0, v0, theta0)
}
\arguments{
\item{v0}{initial velocity in m/s}

\item{y0}{initial height in m}

\item{theta0}{initial angle in degrees}

\item{t}{Time of flight in seconds}
}
\value{
x Displacement in the horizontal direction as a function of time (in meters)

vx speed in the horizontal direction as a function of time (in m/s units)

y Displacement in the vertical direction as a function of time (in meters)

vy speed in the vertical direction as a function of time (in m/s units)

y_x Displacement in the vertical direction as a function of horizontal displacement (in meters)
}
\description{
This function takes in the initial velocity, time of flight,
and returns the distance travlled and speed at a given time t
}

