% Generated by roxygen2 (4.0.1): do not edit by hand
\name{projFrictoinLin}
\alias{projFrictoinLin}
\title{Function to calculate displacement and velocity of a projectile when air friction is introduced.
This function only treats the linear case of velocity dependence.}
\usage{
projFrictoinLin(y0, v0, theta0, t, b, m)
}
\arguments{
\item{v0}{initial velocity in m/s}

\item{y0}{initial height in m}

\item{theta0}{initial angle in degrees}

\item{t}{Time of flight in seconds}

\item{b}{drag coefficient in Newtons.seconds/meters}

\item{m}{mass of object in kg}
}
\value{
x_t Displacement in the horizontal direction as a function of time (in meters)

vx_t speed in the horizontal direction as a function of time (in m/s units)

y_t Displacement in the vertical direction as a function of time (in meters)

vy_t speed in the vertical direction as a function of time (in m/s units)

y_x Displacement in the vertical direction as a function of horizontal displacement (in meters)
}
\description{
This function takes in the initial velocity, time of flight,
drag coefficient, and mass of object and returns the distance
travlled and speed at a given time t
}

