% Generated by roxygen2 (4.0.1): do not edit by hand
\name{disp_vel_lin}
\alias{disp_vel_lin}
\title{Function to calculate displaement and velocity of a projectile when air friction is introduced}
\usage{
disp_vel_lin(Vx0 = 10, t = 10, b = 0.1, m = 0.1)
}
\arguments{
\item{Vx0}{initial velocity in the horizontal direction}

\item{t}{Time of flight in seconds}

\item{b}{drag coefficient}

\item{m}{mass of object in kg}
}
\value{
X Displacement in the horizontal direction

Vx speed in the horizontal direction at a given time t
}
\description{
This function takes in the initial velocity, time of flight,
drag coefficient, and mass of object and returns the distance
distance travlled and speed at a given time t
}

