% Generated by roxygen2 (4.0.1): do not edit by hand
\name{projFrictoinLin}
\alias{projFrictoinLin}
\title{projFrictoinLin}
\usage{
projFrictoinLin(t, y0, v0, theta0, b, m)
}
\arguments{
\item{v0}{initial velocity in m/s}

\item{y0}{initial height in m}

\item{theta0}{initial angle in degrees}

\item{t}{Time of flight in seconds}

\item{b}{drag coefficient in Newtons.seconds/meters}

\item{m}{mass of object in kg}
}
\value{
x Displacement in the horizontal direction as a function of time (in meters)

vx speed in the horizontal direction as a function of time (in m/s units)

y Displacement in the vertical direction as a function of time (in meters)

vy speed in the vertical direction as a function of time (in m/s units)

y_x Displacement in the vertical direction as a function of horizontal displacement (in meters)
}
\description{
Function to calculate displacement and velocity of a projectile when air friction is introduced.
This function only treats the linear case of velocity dependence.
}
\details{
This function takes in the initial velocity, time of flight,
drag coefficient, and mass of object and returns the distance
travlled and speed at a given time t
}
\examples{
projFrictoinLin(t = seq(1,5,0.1), y0 = 10, v0 = 50, theta0 = 30, b = 0.5, m = 5)

## Example of how to call both projectile functions
## and make a projectile path plot

vacuum   <- projectile(t = seq(0,5,0.1), y0 = 10, v0 = 50, theta0 = 30)
friction <- projFrictoinLin(t = seq(0,5,0.1), y0 = 10, v0 = 50, theta0 = 30, b = 0.5, m = 5)

x  <- vacuum$x; y  <- vacuum$y; xf <- friction$x; yf <- friction$y_x

plot(x, y, col="blue", pch=19, ylim=c(min(y, yf), max(y, yf)),
     xlab = "Horizontal Distance Travelled (m)",
     ylab = "Vertical Distance Travelled (m)")
points(xf, yf, col="red", pch=21)
legend(x="topleft", legend = c("No Friction", "Friction"),
    col = c("blue", "red"), pch = c(19, 21))
}
\author{
Aous Abdo
}
\seealso{
\code{\link{projectile}}
}

